requirements.txt
python-telegram-bot==21.6
apscheduler==3.10.4
requests==2.32.3
urllib3==2.2.3
docker==7.1.0
